+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -145,6 +145,16 @@ export class AtomicSearchInterface {
   public connectedCallback() {
     this.i18nPromise = this.initI18n();
     setLoadingFlag(this.store, FirstSearchExecutedFlag);
+    this.updateMobileBreakpoint();
+  }
+
+  private updateMobileBreakpoint() {
+    const breakpoint = this.host.querySelector(
+      'atomic-search-layout'
+    )?.mobileBreakpoint;
+    if (breakpoint) {
+      this.store.set('mobileBreakpoint', breakpoint);
+    }
   }
 
   @Watch('searchHub')