+++ packages/atomic/src/components/result-lists/result-list-common.tsx
@@ -12,7 +12,10 @@ import {
   ResultListProps,
   EcommerceDefaultFieldsToInclude,
 } from '@coveo/headless';
-import {Bindings} from '../../utils/initialization-utils';
+import {
+  Bindings,
+  InitializableComponent,
+} from '../../utils/initialization-utils';
 import {
   ResultDisplayDensity,
   ResultDisplayImageSize,
@@ -29,6 +32,24 @@ import {ListDisplayResultsPlaceholder} from './list-display-results-placeholder'
 import {once} from '../../utils/utils';
 import {updateBreakpoints} from '../../utils/replace-breakpoint';
 import {isAppLoaded, setLoadingFlag} from '../../utils/store';
+
+export interface BaseResultList<T = FoldedResult | Result>
+  extends InitializableComponent {
+  host: HTMLElement;
+  templateHasError: boolean;
+  resultListCommon: ResultListCommon;
+
+  setRenderFunction?: SetRenderFunction<T>;
+
+  density?: ResultDisplayDensity;
+  imageSize?: ResultDisplayImageSize;
+  image?: ResultDisplayImageSize;
+  display?: ResultDisplayLayout;
+}
+
+export type SetRenderFunction<T = FoldedResult | Result> = (
+  render: (result: T) => HTMLElement
+) => Promise<void>;
 interface DisplayOptions {
   density: ResultDisplayDensity;
   imageSize?: ResultDisplayImageSize;