+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.70.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.70.0...@coveo/atomic@1.70.1) (2022-06-09)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.70.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.69.0...@coveo/atomic@1.70.0) (2022-06-09)
+
+
+### Bug Fixes
+
+* **atomic:** add max width to status icons (no results & query error) ([#2070](https://github.com/coveo/ui-kit/issues/2070)) ([620d127](https://github.com/coveo/ui-kit/commit/620d12733ff61b3484d5e835c890e6934d66e487))
+
+
+### Features
+
+* **atomic:** add result templates to instant results ([#2061](https://github.com/coveo/ui-kit/issues/2061)) ([968acd2](https://github.com/coveo/ui-kit/commit/968acd27dd5e25435a41208d88294166e4117aed))
+* **atomic:** extract result list interface ([#2064](https://github.com/coveo/ui-kit/issues/2064)) ([4e0054d](https://github.com/coveo/ui-kit/commit/4e0054d4229fe594b8ab44954cf6941cf3745336))
+
+
+
+
+
 # [1.69.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.68.1...@coveo/atomic@1.69.0) (2022-06-08)