+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.57.1](https://github.com/coveo/ui-kit/compare/headless-react@0.57.0...headless-react@0.57.1) (2022-06-09)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.57.0](https://github.com/coveo/ui-kit/compare/headless-react@0.56.2...headless-react@0.57.0) (2022-06-09)
+
+
+### Features
+
+* **atomic:** extract result list interface ([#2064](https://github.com/coveo/ui-kit/issues/2064)) ([4e0054d](https://github.com/coveo/ui-kit/commit/4e0054d4229fe594b8ab44954cf6941cf3745336))
+
+
+
+
+
 ## [0.56.2](https://github.com/coveo/ui-kit/compare/headless-react@0.56.1...headless-react@0.56.2) (2022-06-08)